fotolibros-agno-backend/requirements.txt
# AGNO Framework
agno>=2.3.0
python-dotenv>=1.0.0

# Database
better-sqlite3>=11.0.0  # Para compatibilidad con el proyecto existente

# Utilidades
pillow>=10.0.0  # Para manejo de imágenes
